Fix for newer versions of Rails where relations have local proxy classes
This is instead of being direct instances of ActiveRecord::Relation.

This also fixes a minor issue where the second "include_private" boolean
parameter is passed into "respond_to_missing?". This was causing issues
in the new unit tests which perform is_a checks on the
CinnamonSerial::Base instances.

Also, update Travis to use the latest Ruby 2.6.